// src/types.ts
// App state
import {AxiosStatic} from "axios";

export interface IAppState {
    recipesOverview: IRecipeOverview
    recipeForm: IRecipeForm,
    userSession: IUserSession,
    notifications: INotification[]
}

// User session
export interface IUserSession {
    isAuthenticated: boolean
    userName: string;
}

//
export interface IRecipesPayload<S> {
    $http: AxiosStatic,
    payload: S
}

// Recipes Overview
export interface IRecipeOverview {
    overviewItems: IRecipeOverviewItem[];
    favoriteRecipes: IRecipeOverviewItem[];
}

export interface IRecipeOverviewItem {
    isFavorite?: boolean;
    title?: string;
    ingredients?: string;
    thumbnail?: string;
    href?: string;
    recipeId: number;
    id?: number //autogenerated by json-server
}

// Recipes form
export interface IRecipeForm {
    title: string;
    ingredients: string;
    submitEnabled: boolean;
    [key: string]: any;
}

export interface IPayloadFormHandler {
    field: string;
    value: any;
}

// Notifications
export interface INotification {
    type: AlertType,
    message: string;
    timeout?: number;
}

export enum AlertType {
    NONE,
    SUCCESS = 'success',
    ERROR = 'danger'
}